requirements.txt
Flask
Flask-SQLAlchemy
pyodbc 
pandas
xlrd